Only call syscall.seteuid on linux
