+++ internal/vfs/osfs_fallback.go
@@ -1,0 +1,10 @@
+//go:build !linux
+// +build !linux
+
+package vfs
+
+func (fs *OsFs) setuid() {
+}
+
+func (fs *OsFs) unsetuid() {
+}
